[fix] Separate concrete type with new file
- 새로운 파일로 구체 타입과 추상 타입을 분리
