+++ DrawingApp/DrawingApp/ColorFactory.swift
@@ -12,23 +12,3 @@ protocol ColorFactory {
   
   func make() -> Color?
 }
-
-class RandomColorFactory: ColorFactory {
-  var alphaFactory: AlphaFactory
-  
-  init(alphaFactory: AlphaFactory = RandomAlphaFactory()) {
-    self.alphaFactory = alphaFactory
-  }
-  
-  private func generateRandomFloat() -> Float {
-    Float.random(in: (0.0 ... 1.0))
-  }
-  
-  func make() -> Color? {
-    let r = generateRandomFloat()
-    let g = generateRandomFloat()
-    let b = generateRandomFloat()
-    guard let alpha = alphaFactory.make() else { return nil }
-    return .init(r: r, g: g, b: b, alpha: alpha)
-  }
-}
